+++ android/src/main/java/io/openim/flutter_openim_sdk/manager/IMManager.java
@@ -36,30 +36,4 @@ public class IMManager extends BaseManager {
     public void getLoginStatus(MethodCall methodCall, MethodChannel.Result result) {
         CommonUtil.runMainThreadReturn(result, Open_im_sdk.getLoginStatus());
     }
-
-//    public void getLoginUid(MethodCall methodCall, MethodChannel.Result result) {
-//        CommonUtil.runMainThreadReturn(result, Open_im_sdk.getLoginUid());
-//    }
-
-//    public void getUsersInfo(MethodCall methodCall, MethodChannel.Result result) {
-//        Open_im_sdk.getUsersInfo(
-//                jsonValue(methodCall, "uidList"), new BaseListener(result));
-//    }
-
-//    public void setSelfInfo(MethodCall methodCall, MethodChannel.Result result) {
-//        Open_im_sdk.setSelfInfo(
-//                jsonValue(methodCall), new BaseListener(result));
-//    }
-
-//    public void forceSyncLoginUerInfo(MethodCall methodCall, MethodChannel.Result result) {
-//        Open_im_sdk.forceSyncLoginUerInfo();
-//    }
-
-//    public void forceReConn(MethodCall methodCall, MethodChannel.Result result) {
-//        Open_im_sdk.forceReConn();
-//    }
-
-//    public void setSdkLog(MethodCall methodCall, MethodChannel.Result result) {
-//        Open_im_sdk.setSdkLog(value(methodCall, "sdkLog"));
-//    }
 }